requirements.txt
article==0.1.1
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==5.0.7
django-ckeditor==6.2.0
django-crispy-forms==2.3
django-js-asset==2.2.0
django-modeltranslation==0.19.7
django-phonenumber-field==8.0.0
django-rosetta==0.10.0
Faker==26.1.0
idna==3.7
phonenumbers==8.13.42
pillow==10.4.0
polib==1.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2

